[mdlk] Fix wrong mapping on non-pagezero files
This commit is huge and probably one of the most important in the goal
of finally loading a Mach-O binary.

It fixes the issue where we would point to the incorrect information
in the bind and rebase instructions.

This commit fixes said issue by handling non-pagezero files differently
than files that do have pagezero.

Is recommend to read the comment to get more information